refactor: rename CaseConstrApply
Add comments, better tests
